backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Image Processing & Face Recognition
pillow==10.2.0
opencv-python-headless==4.12.0.88
numpy==2.0.2
scipy==1.13.1

# Utilities
python-dateutil==2.8.2
